Move JsonType to presto-spi
