+++ presto-thrift-connector-api/src/test/java/com/facebook/presto/connector/thrift/api/valuesets/TestPrestoThriftEquatableValueSet.java
@@ -20,7 +20,7 @@ import org.testng.annotations.Test;
 
 import static com.facebook.presto.connector.thrift.api.PrestoThriftBlock.jsonData;
 import static com.facebook.presto.connector.thrift.api.valuesets.PrestoThriftValueSet.fromValueSet;
-import static com.facebook.presto.type.JsonType.JSON;
+import static com.facebook.presto.spi.type.JsonType.JSON;
 import static io.airlift.slice.Slices.utf8Slice;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.testng.Assert.assertEquals;
